# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "hongker",
            "url": "http://hongker.github.io",
            "email": "xiaok2013@live.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/user/auth": {
            "post": {
                "description": "通过邮箱和密码登录，换取token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "邮箱",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "pass",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "通过邮箱和密码注册账户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "success"
                    },
                    "500": {
                        "description": "error"
                    }
                }
            }
        }
    },
    "definitions": {
        "request.UserRegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "pass"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "pass": {
                    "description": "密码",
                    "type": "string"
                }
            }
        }
    }
}